unica: config: move super group name definition to source config
Since it is the same for all devices with the same SSI define it
in the SSI config file. This also allows device maintainers to specify
a group name different than the SSI's one in case the device requires it.

This does not break existing device configs. It helps future development
and opens new use cases making UN1CA build system even more flexible
and universal.